+++ RELEASE_NOTES.md
@@ -127,6 +127,9 @@ compiler/runtime.
 * ADDED: Bash shell script to create a SICStus Prolog `.po` file with the
 Logtalk compiler/runtime.
 
+* ADDED: Bash shell script to create a XSB `.xwam` file with the Logtalk
+compiler/runtime.
+
 * UPDATED: The Windows installation script to also detect ECLiPSe 7.1
 development versions now that the 7.0 version is final.